+++ Blackboard/Core/Nodes/Bases/NaryValue.cs
@@ -1,9 +1,9 @@
 ﻿using Blackboard.Core.Data.Interfaces;
 using Blackboard.Core.Extensions;
+using Blackboard.Core.Nodes.Collections;
 using Blackboard.Core.Nodes.Functions;
 using Blackboard.Core.Nodes.Interfaces;
 using System.Collections.Generic;
-using System.Linq;
 using S = System;
 
 namespace Blackboard.Core.Nodes.Bases {
@@ -60,25 +60,8 @@ namespace Blackboard.Core.Nodes.Bases {
         public bool RemoveParents(IEnumerable<IValueParent<TIn>> parents) =>
             this.sources.RemoveParents(this, parents);
 
-        /// <summary>This removes all the parents from this node.</summary>
-        public void ClearParents() => this.sources.Clear();
-
         /// <summary>The set of parent nodes to this node in the graph.</summary>
-        public IEnumerable<IParent> Parents => this.sources;
-
-        /// <summary>This replaces all instances of the given old parent with the given new parent.</summary>
-        /// <param name="oldParent">The old parent to find all instances with.</param>
-        /// <param name="newParent">The new parent to replace each instance with.</param>
-        /// <returns>True if any parent was replaced, false if that old parent wasn't found.</returns>
-        public bool ReplaceParent(IParent oldParent, IParent newParent) =>
-            this.sources.ReplaceParents(this, oldParent, newParent);
-
-        /// <summary>This will attempt to set all the parents in a node.</summary>
-        /// <remarks>This will throw an exception if there isn't the correct types.</remarks>
-        /// <param name="newParents">The parents to set.</param>
-        /// <returns>True if any parents changed, false if they were all the same.</returns>
-        public bool SetAllParents(List<IParent> newParents) =>
-            this.sources.SetAllParents(this, newParents);
+        public IParentCollection Parents => new VarParents<IValueParent<TIn>>(this, this.sources);
 
         /// <summary>This handles updating this node's value given the parents' values during evaluation.</summary>
         /// <remarks>Any null parents are ignored.</remarks>
