+++ Blackboard/Core/Nodes/Bases/Select.cs
@@ -1,9 +1,8 @@
 ﻿using Blackboard.Core.Data.Caps;
 using Blackboard.Core.Inspect;
-using Blackboard.Core.Extensions;
+using Blackboard.Core.Nodes.Collections;
 using Blackboard.Core.Nodes.Functions;
 using Blackboard.Core.Nodes.Interfaces;
-using System.Collections.Generic;
 using S = System;
 
 namespace Blackboard.Core.Nodes.Bases {
@@ -58,27 +57,10 @@ namespace Blackboard.Core.Nodes.Bases {
         }
 
         /// <summary>The set of parent nodes to this node in the graph.</summary>
-        public IEnumerable<IParent> Parents => IChild.EnumerateParents(this.source1, this.source2, this.source3);
-
-        /// <summary>This replaces all instances of the given old parent with the given new parent.</summary>
-        /// <param name="oldParent">The old parent to find all instances with.</param>
-        /// <param name="newParent">The new parent to replace each instance with.</param>
-        /// <returns>True if any parent was replaced, false if that old parent wasn't found.</returns>
-        public bool ReplaceParent(IParent oldParent, IParent newParent) =>
-            IChild.ReplaceParent(this, ref this.source1, oldParent, newParent) |
-            IChild.ReplaceParent(this, ref this.source2, oldParent, newParent) |
-            IChild.ReplaceParent(this, ref this.source3, oldParent, newParent);
-
-        /// <summary>This will attempt to set all the parents in a node.</summary>
-        /// <remarks>This will throw an exception if there isn't the correct count or types.</remarks>
-        /// <param name="newParents">The parents to set.</param>
-        /// <returns>True if any parents changed, false if they were all the same.</returns>
-        public bool SetAllParents(List<IParent> newParents) {
-            IChild.CheckParentsBeingSet(newParents, false, typeof(IValueParent<Bool>), typeof(T), typeof(T));
-            return IChild.SetParent(this, ref this.source1, newParents[0] as IValueParent<Bool>) |
-                   IChild.SetParent(this, ref this.source2, newParents[1] as T) |
-                   IChild.SetParent(this, ref this.source3, newParents[2] as T);
-        }
+        public IParentCollection Parents => new FixedParents(this).
+            With(() => this.source1, (IValueParent<Bool> parent) => this.source1 = parent).
+            With(() => this.source2, (T parent) => this.source2 = parent).
+            With(() => this.source3, (T parent) => this.source3 = parent);
 
         /// <summary>The node which is currently selected.</summary>
         public T Selected { get; private set; }
