+++ Blackboard/Core/Nodes/Collections/IParentCollection.cs
@@ -1,0 +1,37 @@
+﻿using Blackboard.Core.Nodes.Interfaces;
+using System.Collections.Generic;
+using S = System;
+
+namespace Blackboard.Core.Nodes.Collections {
+
+    /// <summary>This is a collection of parent nodes for an IChild.</summary>
+    public interface IParentCollection {
+        
+        /// <summary>The set of type that each parent could have in this collection.</summary>
+        /// <remarks>For N-ary nodes this can be very long so use this with zip or limit the enumerations.</remarks>
+        public IEnumerable<S.Type> Types { get; }
+
+        /// <summary>The set of parent nodes to this node.</summary>
+        /// <remarks>
+        /// This should not contain null parents, but it might contain repeat parents.
+        /// For example, if a number is the sum of itself (x + x), then the Sum node will return the 'x' parent twice.
+        /// </remarks>
+        public IEnumerable<IParent> Nodes { get; }
+
+        /// <summary>This replaces all instances of the given old parent with the given new parent.</summary>
+        /// <remarks>
+        /// The new parent must be able to take the place of the old parent,
+        /// otherwise this will throw an exception when attempting the replacement of the old parent.
+        /// </remarks>
+        /// <param name="oldParent">The old parent to find all instances with.</param>
+        /// <param name="newParent">The new parent to replace each instance with.</param>
+        /// <returns>True if any parent was replaced, false if that old parent wasn't found.</returns>
+        public bool ReplaceParent(IParent oldParent, IParent newParent);
+
+        /// <summary>This will attempt to set all the parents in a node.</summary>
+        /// <remarks>This will throw an exception if there isn't the correct count or types.</remarks>
+        /// <param name="newParents">The parents to set.</param>
+        /// <returns>True if any parents changed, false if they were all the same.</returns>
+        public bool SetAllParents(List<IParent> newParents);
+    }
+}
