+++ Blackboard/Core/Nodes/Inner/OnChange.cs
@@ -1,5 +1,6 @@
 ﻿using Blackboard.Core.Extensions;
 using Blackboard.Core.Nodes.Bases;
+using Blackboard.Core.Nodes.Collections;
 using Blackboard.Core.Nodes.Functions;
 using Blackboard.Core.Nodes.Interfaces;
 using System.Collections.Generic;
@@ -43,21 +44,7 @@ namespace Blackboard.Core.Nodes.Inner {
             this.sources.RemoveParents(this, parents);
 
         /// <summary>The set of parent nodes to this node in the graph.</summary>
-        public IEnumerable<IParent> Parents => this.sources;
-
-        /// <summary>This replaces all instances of the given old parent with the given new parent.</summary>
-        /// <param name="oldParent">The old parent to find all instances with.</param>
-        /// <param name="newParent">The new parent to replace each instance with.</param>
-        /// <returns>True if any parent was replaced, false if that old parent wasn't found.</returns>
-        public bool ReplaceParent(IParent oldParent, IParent newParent) =>
-            this.sources.ReplaceParents(this, oldParent, newParent);
-
-        /// <summary>This will attempt to set all the parents in a node.</summary>
-        /// <remarks>This will throw an exception if there isn't the correct types.</remarks>
-        /// <param name="newParents">The parents to set.</param>
-        /// <returns>True if any parents changed, false if they were all the same.</returns>
-        public bool SetAllParents(List<IParent> newParents) =>
-            this.sources.SetAllParents(this, newParents);
+        public IParentCollection Parents => new VarParents<IParent>(this, this.sources);
 
         /// <summary>This updates the trigger during an evaluation.</summary>
         /// <returns>This always returns true so that any parent change will trigger this node.</returns>
