+++ Blackboard/Core/Nodes/Outer/Toggler.cs
@@ -1,5 +1,6 @@
 ﻿using Blackboard.Core.Data.Caps;
 using Blackboard.Core.Nodes.Bases;
+using Blackboard.Core.Nodes.Collections;
 using Blackboard.Core.Nodes.Functions;
 using Blackboard.Core.Nodes.Interfaces;
 using System.Collections.Generic;
@@ -58,27 +59,10 @@ namespace Blackboard.Core.Nodes.Outer {
         }
 
         /// <summary>The set of parent nodes to this node in the graph.</summary>
-        public IEnumerable<IParent> Parents => IChild.EnumerateParents(this.toggle, this.reset, this.resetValue);
-
-        /// <summary>This replaces all instances of the given old parent with the given new parent.</summary>
-        /// <param name="oldParent">The old parent to find all instances with.</param>
-        /// <param name="newParent">The new parent to replace each instance with.</param>
-        /// <returns>True if any parent was replaced, false if that old parent wasn't found.</returns>
-        public bool ReplaceParent(IParent oldParent, IParent newParent) =>
-            IChild.ReplaceParent(this, ref this.toggle,     oldParent, newParent) |
-            IChild.ReplaceParent(this, ref this.reset,      oldParent, newParent) |
-            IChild.ReplaceParent(this, ref this.resetValue, oldParent, newParent);
-
-        /// <summary>This will attempt to set all the parents in a node.</summary>
-        /// <remarks>This will throw an exception if there isn't the correct count or types.</remarks>
-        /// <param name="newParents">The parents to set.</param>
-        /// <returns>True if any parents changed, false if they were all the same.</returns>
-        public bool SetAllParents(List<IParent> newParents) {
-            IChild.CheckParentsBeingSet(newParents, false, typeof(ITriggerParent), typeof(ITriggerParent), typeof(IValueParent<Bool>));
-            return IChild.SetParent(this, ref this.toggle,     newParents[0] as ITriggerParent) |
-                   IChild.SetParent(this, ref this.reset,      newParents[1] as ITriggerParent) |
-                   IChild.SetParent(this, ref this.resetValue, newParents[2] as IValueParent<Bool>);
-        }
+        public IParentCollection Parents => new FixedParents(this).
+            With(() => this.toggle,     (ITriggerParent     parent) => this.toggle     = parent).
+            With(() => this.reset,      (ITriggerParent     parent) => this.reset      = parent).
+            With(() => this.resetValue, (IValueParent<Bool> parent) => this.resetValue = parent);
 
         /// <summary>This will determine the new value the node should be set to.</summary>
         /// <returns>The new value that the node should be set to.</returns>
