+++ Blackboard/Parser/Optimization/Rules/Coalescer.cs
@@ -15,7 +15,7 @@ namespace Blackboard.Parser.Optimization.Rules {
         /// </summary>
         /// <param name="node">The node to incorporate parents into.</param>
         static private void incorporateParents(ICoalescable node) {
-            List<IParent> parents = node.Parents.ToList();
+            List<IParent> parents = node.Parents.Nodes.ToList();
             bool changed = false;
             for (int i = 0; i < parents.Count; ++i) {
                 IParent parent = parents[i];
@@ -27,8 +27,8 @@ namespace Blackboard.Parser.Optimization.Rules {
                     // Incorporate the parent's parents here.
                     parents.RemoveAt(i);
                     if (parent is IChild childParent) {
-                        parents.InsertRange(i, childParent.Parents);
-                        foreach (IParent parentsParent in childParent.Parents) {
+                        parents.InsertRange(i, childParent.Parents.Nodes);
+                        foreach (IParent parentsParent in childParent.Parents.Nodes) {
                             bool removed = parentsParent?.RemoveChildren(childParent) ?? false;
                             if (removed) parentsParent?.AddChildren(node);
                         }
@@ -39,7 +39,7 @@ namespace Blackboard.Parser.Optimization.Rules {
             }
 
             if (changed)
-                node.SetAllParents(parents);
+                node.Parents.SetAllParents(parents);
         }
 
         /// <summary></summary>
@@ -84,7 +84,7 @@ namespace Blackboard.Parser.Optimization.Rules {
                     args.Root = newNode;
                 else if (newNode is IParent newParent && cNode is IParent cParent ) {
                     foreach (IChild child in cParent.Children.ToList())
-                        child.ReplaceParent(cParent, newParent);
+                        child.Parents.ReplaceParent(cParent, newParent);
                 }
             }
         }
